refactored all mirror/msms charts to mirrorchartutils; renamed mirrorchartutils to MirrorChartFactory; removed duplicate code in MirrorScanWindowFX
