+++ src/main/java/io/github/mzmine/modules/visualization/spectra/simplespectra/mirrorspectra/MirrorScanWindow.java
@@ -17,6 +17,7 @@
  */
 package io.github.mzmine.modules.visualization.spectra.simplespectra.mirrorspectra;
 
+import io.github.mzmine.util.MirrorChartFactory;
 import io.github.mzmine.util.color.SimpleColorPalette;
 import java.awt.BorderLayout;
 import java.awt.Color;
@@ -36,11 +37,8 @@ import io.github.mzmine.datamodel.DataPoint;
 import io.github.mzmine.datamodel.Scan;
 import io.github.mzmine.gui.chartbasics.gui.swing.EChartPanel;
 import io.github.mzmine.main.MZmineCore;
-import io.github.mzmine.modules.visualization.spectra.multimsms.SpectrumChartFactory;
 import io.github.mzmine.modules.visualization.spectra.multimsms.pseudospectra.PseudoSpectraRenderer;
 import io.github.mzmine.modules.visualization.spectra.multimsms.pseudospectra.PseudoSpectrumDataSet;
-import io.github.mzmine.util.color.Colors;
-import io.github.mzmine.util.color.Vision;
 import io.github.mzmine.util.spectraldb.entry.DBEntryField;
 import io.github.mzmine.util.spectraldb.entry.DataPointsTag;
 import io.github.mzmine.util.spectraldb.entry.SpectralDBPeakIdentity;
@@ -76,7 +74,7 @@ public class MirrorScanWindow extends JFrame {
   public void setScans(String labelA, double precursorMZA, double rtA, DataPoint[] dpsA,
       String labelB, double precursorMZB, double rtB, DataPoint[] dpsB) {
     contentPane.removeAll();
-    mirrorSpecrumPlot = SpectrumChartFactory.createMirrorChartPanel(labelA, precursorMZA, rtA, dpsA,
+    mirrorSpecrumPlot = MirrorChartFactory.createMirrorChartPanel(labelA, precursorMZA, rtA, dpsA,
         labelB, precursorMZB, rtB, dpsB, false, true);
     contentPane.add(mirrorSpecrumPlot, BorderLayout.CENTER);
     contentPane.revalidate();
@@ -91,7 +89,7 @@ public class MirrorScanWindow extends JFrame {
    */
   public void setScans(Scan scan, Scan mirror) {
     contentPane.removeAll();
-    mirrorSpecrumPlot = SpectrumChartFactory.createMirrorChartPanel(scan, mirror,
+    mirrorSpecrumPlot = MirrorChartFactory.createMirrorChartPanel(scan, mirror,
         scan.getScanDefinition(), mirror.getScanDefinition(), false, true);
     contentPane.add(mirrorSpecrumPlot, BorderLayout.CENTER);
     contentPane.revalidate();
@@ -102,7 +100,7 @@ public class MirrorScanWindow extends JFrame {
   public void setScans(Scan scan, Scan mirror, String labelA, String labelB) {
     contentPane.removeAll();
     mirrorSpecrumPlot =
-        SpectrumChartFactory.createMirrorChartPanel(scan, mirror, labelA, labelB, false, true);
+        MirrorChartFactory.createMirrorChartPanel(scan, mirror, labelA, labelB, false, true);
     contentPane.add(mirrorSpecrumPlot, BorderLayout.CENTER);
     contentPane.revalidate();
     contentPane.repaint();
@@ -111,7 +109,6 @@ public class MirrorScanWindow extends JFrame {
   /**
    * Based on a data base match to a spectral library
    *
-   * @param row
    * @param db
    */
   public void setScans(SpectralDBPeakIdentity db) {
@@ -151,7 +148,7 @@ public class MirrorScanWindow extends JFrame {
 
     contentPane.removeAll();
     // create without data
-    mirrorSpecrumPlot = SpectrumChartFactory.createMirrorChartPanel(
+    mirrorSpecrumPlot = MirrorChartFactory.createMirrorChartPanel(
         "Query: " + scan.getScanDefinition(), precursorMZA, rtA, null, "Library: " + db.getName(),
         precursorMZB == null ? 0 : precursorMZB, rtB, null, false, true);
     mirrorSpecrumPlot.setMaximumDrawWidth(4200);
